Refactor login router name
