+++ back-end/src/routes/index.ts
@@ -1,1 +1,1 @@
-export * from "./LoginRouter";
+export * from "./loginRouter";
